pyproject.toml
[project]
name = "detect-ccd-change"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.14"
dependencies = [
    "beautifulsoup4>=4.14.3",
    "httpx>=0.28.1",
    "python-dotenv>=1.2.2",
]
